updated the location of the install --prefix

+++ Bash/Installer Scripts/GitHub Projects/build-libpng.sh
@@ -3,25 +3,26 @@
 
 ##  Github Script: https://github.com/slyfox1186/script-repo/blob/main/Bash/Installer%20Scripts/GitHub%20Projects/build-libpng.sh
 ##  Purpose: Build GNU libpng
-##  Updated: 08.31.23
-##  Script version: 1.1
+##  Updated: 07.03.23
+##  Script version: 1.2
 
-if [ "${EUID}" -eq '0' ]; then
+if [[ "$EUID" -eq 0 ]]; then
     echo "You must run this script without root or sudo."
     exit 1
 fi
 
 # Set variables
-script_ver=1.1
-cwd="$PWD"/libpng-build-script
-install_dir=/usr/local
-user_agent='Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/121.0.0.0 Safari/537.36'
-web_repo=https://github.com/slyfox1186/script-repo
+script_ver="1.2"
+prog_name="libpng"
+cwd="$PWD/libpng-build-script"
+version=$(curl -fsS "https://github.com/pnggroup/libpng/tags/" | grep -oP '/tag/v\K\d+\.\d+\.\d+' | head -n1)
+archive_name="$prog_name-$version"
+install_dir="/usr/local/programs/$archive_name"
 debug=OFF
 
-printf "\n%s\n%s\n\n" \
-    "libpng build script - v${script_ver}" \
-    '==============================================='
+echo "libpng build script - v${script_ver}"
+echo "==============================================="
+echo
 
 # Create output directory
 
@@ -42,50 +43,46 @@ export CC CXX CFLAGS CXXFLAGS PATH PKG_CONFIG_PATH
 # Create functions
 
 exit_function() {
-    printf "\n%s\n\n%s\n\n" \
-        'Make sure to star this repository to show your support!' \
-        "$web_repo"
+    echo "Make sure to star this repository to show your support!"
+    echo "https://github.com/slyfox1186/script-repo"
+    echo
     exit 0
 }
 
 fail_fn() {
-    printf "\n%s\n\n%s\n\n" \
-        "$1" \
-        "To report a bug create an issue at: $web_repo/issues"
+    echo "$1"
+    echo "To report a bug create an issue at: https://github.com/slyfox1186/script-repo/issues"
+    echo
     exit 1
 }
 
 cleanup_fn() {
     local choice
 
-    printf "%s\n%s\n%s\n\n%s\n%s\n\n" \
-        '============================================' \
-        '  Do you want to clean up the build files?  ' \
-        '============================================' \
-        '[1] Yes' \
-        '[2] No'
-    read -p 'Your choices are (1 or 2): ' choice
+    echo "============================================"
+    echo "  Do you want to clean up the build files?  "
+    echo "============================================"
+    echo "[1] Yes"
+    echo "[2] No"
+    echo
+    read -p "Your choices are (1 or 2): " choice
 
     case "${choice}" in
         1)      sudo rm -fr "$cwd";;
         2)      echo;;
         *)
-                clear
-                printf "%s\n\n" 'Bad user input. Reverting script...'
-                sleep 3
                 unset choice
-                clear
+                echo
                 cleanup_fn
                 ;;
     esac
 }
 
 build() {
-    printf "\n%s\n%s\n" \
-        "building $1 - version $2" \
-        '===================================='
+    echo "building $1 - version $2"
+    echo "===================================="
 
-    if [ -f "$cwd/$1.done" ]; then
+    if [[ -f "$cwd/$1.done" ]]; then
         if grep -Fx "$2" "$cwd/$1.done" >/dev/null; then
             echo "$1 version $2 already built. Remove $cwd/$1.done lockfile to rebuild it."
             return 1
@@ -101,17 +98,17 @@ build() {
 }
 
 execute() {
-    echo "$ ${*}"
+    echo "$ $*"
 
-    if [ "${debug}" = 'ON' ]; then
+    if [[ "${debug}" = "ON" ]]; then
         if ! output=$("$@"); then
-            notify-send 5000 "Failed to execute: ${*}"
-            fail_fn "Failed to execute: ${*}"
+            notify-send 5000 "Failed to execute: $*"
+            fail_fn "Failed to execute: $*"
         fi
     else
         if ! output=$("$@" 2>&1); then
-            notify-send 5000 "Failed to execute: ${*}"
-            fail_fn "Failed to execute: ${*}"
+            notify-send 5000 "Failed to execute: $*"
+            fail_fn "Failed to execute: $*"
         fi
     fi
 }
@@ -131,34 +128,31 @@ download() {
     target_file="$dl_path/$dl_file"
     target_dir="$dl_path/$output_dir"
 
-    if [ -f "${target_file}" ]; then
+    if [[ -f "$target_file" ]]; then
         echo "The file \"$dl_file\" is already downloaded."
     else
-        echo "Downloading \"${dl_url}\" saving as \"$dl_file\""
-        if ! curl -A "$user_agent" -Lso "${target_file}" "${dl_url}"; then
+        echo "Downloading \"$dl_url\" saving as \"$dl_file\""
+        if ! curl -Lso "$target_file" "$dl_url"; then
             printf "\n%s\n\n" "The script failed to download \"$dl_file\" and will try again in 10 seconds..."
             sleep 10
-            if ! curl -A "$user_agent" -Lso "${target_file}" "${dl_url}"; then
+            if ! curl -Lso "$target_file" "$dl_url"; then
                 fail_fn "The script failed to download \"$dl_file\" twice and will now exit:Line ${LINENO}"
             fi
         fi
-        echo 'Download Completed'
+        echo "Download Completed"
     fi
 
-    if [ -d "$target_dir" ]; then
-        sudo rm -fr "$target_dir"
-    fi
-
+    [[ -d "$target_dir" ]] && sudo rm -fr "$target_dir"
     mkdir -p "$target_dir"
 
-    if [ -n "$3" ]; then
-        if ! tar -xf "${target_file}" -C "$target_dir" 2>/dev/null >/dev/null; then
-            sudo rm "${target_file}"
+    if [[ -n "$3" ]]; then
+        if ! tar -xf "$target_file" -C "$target_dir" 2>/dev/null >/dev/null; then
+            sudo rm "$target_file"
             fail_fn "The script failed to extract \"$dl_file\" so it was deleted. Please re-run the script:Line ${LINENO}"
         fi
     else
-        if ! tar -xf "${target_file}" -C "$target_dir" --strip-components 1 2>/dev/null >/dev/null; then
-            sudo rm "${target_file}"
+        if ! tar -xf "$target_file" -C "$target_dir" --strip-components 1 2>/dev/null >/dev/null; then
+            sudo rm "$target_file"
             fail_fn "The script failed to extract \"$dl_file\" so it was deleted. Please re-run the script:Line ${LINENO}"
         fi
     fi
@@ -177,14 +171,14 @@ pkgs_fn() {
 
     for i in ${pkgs[@]}
     do
-        missing_pkg="$(sudo dpkg -l | grep -o "${i}")"
+        missing_pkg="$(sudo dpkg -l | grep -q "$i")"
     
-        if [ -z "${missing_pkg}" ]; then
-            missing_pkgs+=" ${i}"
+        if [[ -z "$missing_pkg" ]]; then
+            missing_pkgs+=" $i"
         fi
     done
     
-    if [ -n "$missing_pkgs" ]; then
+    if [[ -n "$missing_pkgs" ]]; then
         sudo apt install $missing_pkgs
         sudo apt -y autoremove
         clear
@@ -198,25 +192,21 @@ build_done() { echo "$2" > "$cwd/$1.done"; }
 pkgs_fn
 
 # Build program from source
-if build "libpng" "1.6.40"; then
-    download "https://github.com/glennrp/libpng/archive/refs/tags/v1.6.40.tar.gz" "libpng-1.6.40.tar.gz"
+if build "libpng" "$version"; then
+    download "https://github.com/pnggroup/libpng/archive/refs/tags/v$version.tar.gz" "$archive_name.tar.gz"
     execute autoupdate
     execute autoreconf -fi
     execute ./configure --prefix="$install_dir" \
-                         --{build,host}=x86_64-linux-gnu \
                          --disable-shared \
                          --enable-hardware-optimizations \
                          --enable-unversioned-links \
                          --with-binconfigs \
                          --with-pic \
-                         --with-pkgconfigdir="${PKG_CONFIG_PATH}" \
+                         --with-pkgconfigdir="$PKG_CONFIG_PATH" \
                          --with-zlib-prefix
     execute make "-j$(nproc --all)"
-    execute sudo make install-header-links
-    execute sudo make install-library-links
     execute sudo make install
-    execute make distclean
-    build_done "libpng" "1.6.40"
+    build_done "libpng" "$version"
 fi
 
 # Prompt user to clean up files
